It's a fucking refactor to add a entry validation
Now all the data sent is sent as string, including the amount.
This code is a shit. I'll refactor all od this.

+++ models/entry.go
@@ -5,7 +5,7 @@ import "time"
 type Entry struct {
 	Id      string
 	Name    string
-	Amount  float32
+	Amount  float64
 	Date    time.Time
 	Comment string
 }
